\name{vcmasf-package}
\alias{vcmasf-package}
\alias{vcmasf}
\docType{package}
\title{
\packageTitle{vcmasf}
}
\description{
\packageDescription{vcmasf}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{vcmasf}
\packageIndices{vcmasf}
}
\author{
\packageAuthor{vcmasf}

Maintainer: \packageMaintainer{vcmasf}
}
\references{
  See paper \href{https://arxiv.org/abs/2201.10063}{Varying Coefficient Model via Adaptive Spline Fitting} for more details.
}
\keyword{ package }

